+++ src/modules/auth/authentication.service.ts
@@ -1,11 +1,12 @@
-import { Injectable, Logger } from "@nestjs/common";
+import { Injectable, Logger, NotImplementedException } from "@nestjs/common";
 import { JwtService } from "@nestjs/jwt";
 import configuration from "../../configuration";
 import { GamevaultUser } from "../users/gamevault-user.entity";
 import { RegisterUserDto } from "../users/models/register-user.dto";
 import { UsersService } from "../users/users.service";
-import { GamevaultJwt } from "./models/gamevault-jwt.interface";
-import { LoginDto } from "./models/login.dto";
+import { GamevaultJwtPayload } from "./models/gamevault-jwt-payload.interface";
+import { RefreshTokenDto } from "./models/refresh-token.dto";
+import { TokenPairDto } from "./models/token-pair.dto";
 
 @Injectable()
 export class AuthenticationService {
@@ -15,12 +16,12 @@ export class AuthenticationService {
     private readonly jwtService: JwtService,
   ) {}
 
-  async login(requestUser: GamevaultUser): Promise<LoginDto> {
+  async login(requestUser: GamevaultUser): Promise<TokenPairDto> {
     const user = await this.usersService.findOneByUsernameOrFail(
       requestUser.username,
     );
 
-    const payload: GamevaultJwt = {
+    const payload: GamevaultJwtPayload = {
       sub: user.id.toString(),
       name: [user.first_name, user.last_name].filter(Boolean).join(" ") || null,
       given_name: user.first_name,
@@ -31,8 +32,7 @@ export class AuthenticationService {
       birthdate: user.birth_date?.toISOString(),
     };
 
-    const loginDto: LoginDto = {
-      id: payload.sub,
+    const loginDto: TokenPairDto = {
       access_token: this.jwtService.sign({ payload }),
       refresh_token: this.jwtService.sign(
         { payload },
@@ -46,23 +46,25 @@ export class AuthenticationService {
     this.logger.debug({
       message: `Issued token for user ${requestUser.username}`,
       token_payload: payload,
-      token_expires_in: configuration.AUTH.ACCESS_TOKEN.EXPIRES_IN,
+      access_token_expires_in: configuration.AUTH.ACCESS_TOKEN.EXPIRES_IN,
+      refresh_token_expires_in: configuration.AUTH.REFRESH_TOKEN.EXPIRES_IN,
     });
     return loginDto;
   }
 
-  async refresh(user: GamevaultUser): Promise<LoginDto> {
+  async refresh(user: GamevaultUser): Promise<TokenPairDto> {
     this.logger.debug(`Refreshing token for user ${user.username}`);
-    const loginDto = await this.login(user);
-    delete loginDto.refresh_token;
-    return loginDto;
+    return await this.login(user);
   }
 
   async register(dto: RegisterUserDto): Promise<GamevaultUser> {
     return this.usersService.register(dto);
   }
 
-  async logout() {
-    /* TODO Implement logout */
+  async revoke(dto: RefreshTokenDto) {
+    //TODO: Implement token revocation
+    throw new NotImplementedException(
+      "Revoking tokens is not implemented yet.",
+    );
   }
 }