+++ src/modules/auth/controllers/authentication.controller.ts
@@ -1,6 +1,14 @@
-import { Controller, Logger, Post, Request, UseGuards } from "@nestjs/common";
+import {
+  Body,
+  Controller,
+  Logger,
+  Post,
+  Request,
+  UseGuards,
+} from "@nestjs/common";
 import {
   ApiBearerAuth,
+  ApiBody,
   ApiOkResponse,
   ApiOperation,
   ApiTags,
@@ -9,7 +17,8 @@ import { SkipGuards } from "../../../decorators/skip-guards.decorator";
 import { GamevaultUser } from "../../users/gamevault-user.entity";
 import { AuthenticationService } from "../authentication.service";
 import { RefreshTokenGuard } from "../guards/refresh-token.guard";
-import { LoginDto } from "../models/login.dto";
+import { RefreshTokenDto } from "../models/refresh-token.dto";
+import { TokenPairDto } from "../models/token-pair.dto";
 
 @Controller("auth")
 @ApiTags("auth")
@@ -27,12 +36,25 @@ export class GamevaultJwtController {
       "This endpoint takes a valid refresh token from the request and issues a new access token. The refresh token must be sent in the request (usually as a cookie or in the authorization header).",
     operationId: "postAuthRefresh",
   })
-  @ApiOkResponse({ type: () => LoginDto })
+  @ApiOkResponse({ type: () => TokenPairDto })
   async postAuthRefresh(
     @Request() req: { user: GamevaultUser },
-  ): Promise<LoginDto> {
+  ): Promise<TokenPairDto> {
     return this.authService.refresh(req.user);
   }
 
-  // TODO: implement revoke
+  @Post("revoke")
+  @ApiBody({ type: () => RefreshTokenDto })
+  @SkipGuards()
+  @ApiOperation({
+    summary: "revokes the refresh token of the user",
+    description:
+      "This endpoint takes a valid refresh token from the request and revokes it. The refresh token must be sent in the request (usually as a cookie or in the authorization header).",
+    operationId: "postAuthRevoke",
+  })
+  async postAuthRevoke(
+    @Body() refreshTokenDto: RefreshTokenDto,
+  ): Promise<void> {
+    return this.authService.revoke(refreshTokenDto);
+  }
 }