+++ src/modules/auth/guards/authentication.guard.ts
@@ -25,11 +25,6 @@ export class AuthenticationGuard extends AuthGuard("auth") {
         >(SKIP_GUARDS_KEY, [context.getHandler(), context.getClass()])
         ?.includes(this.constructor.name)
     ) {
-      this.logger.debug({
-        message: "Skipping Refresh Authentication Checks.",
-        reason: "skip-guards is set to true for this route.",
-        route: context.getHandler(),
-      });
       return true;
     }