+++ src/modules/auth/guards/authorization.guard.ts
@@ -37,11 +37,6 @@ export class AuthorizationGuard implements CanActivate {
         >(SKIP_GUARDS_KEY, [context.getHandler(), context.getClass()])
         ?.includes(this.constructor.name)
     ) {
-      this.logger.debug({
-        message: "Skipping Refresh Authentication Checks.",
-        reason: "skip-guards is set to true for this route.",
-        route: context.getHandler(),
-      });
       return true;
     }