+++ src/modules/auth/guards/basic-auth.guard.ts
@@ -18,11 +18,6 @@ export class BasicAuthGuard extends AuthGuard("basic") {
         >(SKIP_GUARDS_KEY, [context.getHandler(), context.getClass()])
         ?.includes(this.constructor.name)
     ) {
-      this.logger.debug({
-        message: "Skipping Refresh Authentication Checks.",
-        reason: "skip-guards is set to true for this route.",
-        route: context.getHandler(),
-      });
       return true;
     }