+++ src/modules/auth/models/login.dto.ts
@@ -1,8 +1,0 @@
-import { GamevaultJwt } from "./gamevault-jwt.interface";
-
-export class LoginDto {
-  id: string;
-  access_token: string;
-  refresh_token?: string;
-  payload?: GamevaultJwt;
-}