+++ src/modules/auth/models/oauth2-user-profile.interface.ts
@@ -1,8 +1,0 @@
-export default interface Oauth2UserProfile {
-  id?: string;
-  username?: string;
-  email?: string;
-  first_name?: string;
-  last_name?: string;
-  birthdate?: string;
-}